pyproject.toml
[tool.poetry]
name = "document-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
docxtpl = "0.17.0"
pyqrcode = "^1.2.1"
pydantic-settings = "^2.2.1"
requests = "^2.32.3"
latex2mathml = "^3.77.0"
pylint-pydantic = "^0.3.2"
python-docx = "1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
